requirements.txt
python-telegram-bot==20.7
discord.py==2.4.0
aiohttp==3.9.5
aiosqlite==0.20.0
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
stripe==10.12.0
